+++ docs/source/changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.12.6 (2025-XX-XX)
-- Feat: Automatically check cache-validity of polars DataFrame tasks marked as lazy
+- Feat: Automatically check cache-validity of polars and pandas DataFrame tasks marked as lazy
 
 ## 0.12.5 (2025-11-26)
 - Workaround snowflake sqlalchemy dialect to enable ExternalTableReference to other database